move reparam to main module
