+++ test/test_autoguide.py
@@ -21,13 +21,13 @@ from numpyro.infer.autoguide import (
     AutoMultivariateNormal
 )
 from numpyro.nn.auto_reg_nn import AutoregressiveNN
-from numpyro.contrib.reparam import TransformReparam, reparam
 import numpyro.distributions as dist
 from numpyro.distributions import constraints, transforms
 from numpyro.distributions.flows import InverseAutoregressiveTransform
 from numpyro.handlers import substitute
 from numpyro.infer import ELBO, SVI
 from numpyro.infer.initialization import init_to_median
+from numpyro.infer.reparam import TransformReparam
 from numpyro.util import fori_loop
 
 init_strategy = init_to_median(num_samples=2)
@@ -163,7 +163,7 @@ def test_uniform_normal():
 
     def model(data):
         alpha = numpyro.sample('alpha', dist.Uniform(0, 1))
-        with reparam(config={'loc': TransformReparam()}):
+        with numpyro.handlers.reparam(config={'loc': TransformReparam()}):
             loc = numpyro.sample('loc', dist.Uniform(0, alpha))
         numpyro.sample('obs', dist.Normal(loc, 0.1), obs=data)
 
@@ -231,7 +231,7 @@ def test_dynamic_supports():
 
     def actual_model(data):
         alpha = numpyro.sample('alpha', dist.Uniform(0, 1))
-        with reparam(config={'loc': TransformReparam()}):
+        with numpyro.handlers.reparam(config={'loc': TransformReparam()}):
             loc = numpyro.sample('loc', dist.Uniform(0, alpha))
         numpyro.sample('obs', dist.Normal(loc, 0.1), obs=data)
